missive: Capture and nullify INIT_RESUME
These upload calls take place if the queue is not empty at restart,
and in most cases we just want to ignore it, or in simple cases
accept what was written. It is not possible to always set expectations,
because confirmation and corruption remove files differently in
different configurations.

BUG=b:259398986
BUG=b:258913500
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I2d4882d2d0353fc0ec7daab6d142796f932bf2d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4033330
Reviewed-by: Hong Xu <xuhong@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>