cryptohome: Add base class for Full API Behaviour Test in UserDataAuth
The UserDataAuthApi base class is for a set of integration-style unit
tests that is aimed at stressing the negative cases from an API usage
perspective. This differs from other unit tests in which it is written
in more of a integration test style and verifies the behaviour of
cryptohomed APIs rather than the UserDataAuth class.

BUG=b:259636795
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I965ce0ec0bd9f3761facdc9e18932e40547d7ecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4037563
Commit-Queue: John L Chen <zuan@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: John L Chen <zuan@chromium.org>