cryptohome: InstallAttributes receives Platform ptr
Change the InstallAttributes class to receive a Platform pointer
in the constructor, rather than have a separate set_platform()
setter.

This fixes the problem that when constructing UserDataAuth with a
test Platform the default-created InstallAttributes instance was
getting a real Platform. Also it's cleaner to pass dependencies
directly to constructor.

BUG=b:254864841
TEST=cros_workon_make cryptohome --test

Change-Id: Iccdd6d3595bc43648504227b83d0ed7af7c1a37d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4037669
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>