cryptohome: Make AuthFactory assignable
Despite the original intent, the AuthFactor class ended up being a value
object that contains a copy of the identifying and state information for
a factor, rather than an interface which is implemented in different
ways (AuthBlock ended up filling that role).

Given the way it is actually used then, it makes sense for it to be
assignable. This allows it to be used in a lot more places (e.g. a lot
of containers require assignable objects), and it better aligns with
what are generally considered to be good patterns in C++ (copyable but
not copy-assignable objects are considered weird and usually a mistake).

In practice, making this copyable just means de-consting the member
variables.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ia610e4c079e8d65eac3ee614103bc99e24077142
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4039142
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>