vm_tools: Add mkfs_opts and tune2fs_opts to CreateDiskImageRequest
Migration of ARCVM's /data from virtio-fs to virtio-blk requires that
a disk image created by vm_concierge (= the destination of the
migration) to be properly formatted on creation.

To achieve that, this CL adds options for CreateDiskImageRequest to run
mkfs and tune2fs with specified params.

Note that tune2fs is applied to a temporary path that does not contain
the equal sign '=' to work around a path resolution bug (b/267134417).

BUG=b:258278176
BUG=b:267134417
TEST=emerge and cros deploy chromeos-base/system_api, vm_host_tools
TEST=Manually check with CL:4042342 that the disk is properly created
and formatted.

Change-Id: I515fe5e6913c22c373226c30477a1eab5cdd22da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4042301
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>