cryptohome: Fix Platform overriding in UserOldestActivityTimestampManager
Fix UserDataAuth to pass the correct Platform* pointer when
constructing UserOldestActivityTimestampManager. The issue was
specific to tests or fuzzers: the default-constructed real Platform
implementation was passed instead of the one injected via
set_platform(). This meant that this piece of logic was performing
real file system operations in the root folder even in tests or
fuzzers.

BUG=b:254864841
TEST=add asserts into every Platform method, run userdataauth_fuzzer, check there's no assert with UserOldestActivityTimestampManager on stack

Change-Id: I2b49758e075395be8599f3e8e80bacbaea7dd911
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4040932
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>