cryptohome: Tweak StartAuthSession handling in UDA fuzzer
Tweak the special-cased StartAuthSession handling in
userdataauth_fuzzer to additionally check the D-Bus method name.
This significantly reduces the number of false positives (when
the fuzzer would wrongly add a response field as a "breadcrumb",
misinterpreting it as an AuthSessionId), and gives a significant
speedup for the fuzzing coverage in practice.

BUG=b:254864841
TEST=add crash into AuthSessionManager::FindAuthSession() successful exit branch, run userdataauth_fuzzer for a few minutes, check the crash is discovered

Change-Id: I1c7d1577519d11650e047d87200d41f4c8d1e98a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4043181
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>