cryptohome: Add uss_migrator class
UssMigrator creates the UserSecretStash file and adds the
encrypted main key with the migration secret. If there is
an existing UserSecretStash migrator loads the USS
and the main key using the same migration secret.
UssMigrator runs completion_callback when finished. If
user_secret_stash is created or loaded successfully and
wrapped uss_main_key is created & added or obtained from
USS by the migration secret successfully they are passed
to the completion callback, otherwise null and empty
values are passed.

This class doesn't do any cleanup for the corrupted
UserSecretStash or the AuthFactors. It'll be handled separately
in the follow ups.

BUG=b:258465405
TEST=Unit tests pass.

Change-Id: I00f00a03c411f372b8c4cb4587f7a3fd3fcf5839
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4049049
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>