rgbkbd: add tests for ReinitializeOnDeviceReconnected
Add missing tests to help with a future refactoring of
the code that handles Shift highlighting when Caps Lock
is on.

GetRainbowModeColorsWithoutShiftKeysForTesting was changed
to GetRainbowModeColorsWithShiftKeysHighlightedForTesting.
Now it also adds highlighted shift entries at the beginning,
mimicking the effect of ReinitializeOnDeviceReconnected,
to make it more convenient to generate expected logs for
tests.

BUG=b:259589687
TEST=cros_workon_make --board=brya rgbkbd --test

Change-Id: Ic5c94a3756ac8e131abe3d58bc97dac6704a8603
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4044882
Reviewed-by: David Padlipsky <dpad@google.com>
Commit-Queue: Anton Swifton <swifton@google.com>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Tested-by: Anton Swifton <swifton@google.com>