cryptohome: Add a fake of data_migrator::MigrationHelperDelegate
Currently, the unit test of MigrationHelper depends on
DircryptoMigrationHelperDelegate to test if it correctly skips the
migration of some files using the results of a delegate method
ShouldMigrateFile().
This patch removes this dependency by introducing a fake of
MigrationHelperDelegate and using it in MigrationHelper unit test and
rewrite the test to not depend on Ext4 migration-specific logic.
This patch also adds unit tests for DircryptoMigrationHelperDelegate to
check the behavior of ShouldMigrateFile() in the Ext4 migration.

BUG=b:258402655
TEST=Run `FEATURES=test emerge-hatch cryptohome` and check that all
TEST=tests pass and DircryptoMigrationHelperDelegateTest is executed.

Change-Id: I8904751af605b443248a38c4dcc6cde14b1d2a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4055006
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>