typecd: Introduce state for data access Chrome setting
Add a variable to the ChromeFeaturesServiceClient class to track the
Peripheral Data Access setting state.

We'd like to move to a model where Chrome informs typecd of changes to
the setting value, instead of the other way around (to account for
momentary unavailability of the ChromeFeaturesService D-Bus service).
The first step towards accomplishing that is to have a place to maintain
the state of that setting.

This variable isn't used (or even set) anywhere, but that will change
with follow up patches.

BUG=b:260017469
TEST=All unit tests pass.

Change-Id: I36f084449d07f0bbc91577c337bf27869eadc6af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4052761
Reviewed-by: Jameson Thies <jthies@google.com>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>