cryptohome: Expose aeskeylocker info via dbus call
This change adds a new dbus call exposing information if AES Keylocker
is enabled. This information is necessary for hiberman (CrOS
hibernation). In case of platforms with AES Keylocker enabled the
hibernation goes as low as S4 to maintain the key material (in opposite
to S5 for platforms without AES Keylocker).

BUG=b:257536730
TEST=cros_run_unit_tests --board=${BOARD} --packages cryptohome

Change-Id: I75897da0d5eeca3f7368b9b521e25075a6e3d38d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4049649
Commit-Queue: Jakub Młynarczyk <jakubm@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Tested-by: Jakub Młynarczyk <jakubm@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>