rgbkbd: Refactor rainbow handling in the keyboard controller
Instead of representing the rainbow coloring as an array of
LED-color pairs, it will be represented as an array of 4 or 5
zones (depending on keyboard capabilities) and an array of
corresponding colors. Each zone is an array of LED indices.
This design will be reused for the function that allows to set
the color of each zone separately.

BUG=b:259589687
TEST=cros_workon_make --board=brya rgbkbd --test

Change-Id: I8ebd43a018dc555cbe4af5db04df3103bafbbdc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4050464
Commit-Queue: Anton Swifton <swifton@google.com>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Reviewed-by: David Padlipsky <dpad@google.com>
Tested-by: Anton Swifton <swifton@google.com>