fusebox: implement Truncate, Unlink for the client
There was some existing client code for Truncate, but it spoke a vanilla
D-Bus protocol with the Fusebox server (which has not yet implemented
Truncate support). More recently, we prefer protobufs-over-D-Bus, as it
gives us greater flexibility for protocol changes.

This is one half of a multi-repo change. The other half is at
https://crrev.com/c/4054201

BUG=b:255520194,b:255703917
TEST=manual

Change-Id: I91b1fb7beda3ca64feb7ccae736b28222f91e4c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4054240
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>