dlp: Run database on a separate thread.
The database operations include IO file reads/writes
and shouldn't be run on the main thread to not block it.

Another issue that is resolved here is that accessing it's
own database files ends up with another file open request
to the daemon that should always silently be allowed.

BUG=b:259687730
TEST=emerge-{BOARD} chromeos-base/dlp --test

Change-Id: I5dd6f6320ec0a140a1c9709ce828abe0632a1cfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4055605
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>