cryptohome: Use Platform when checking aeskl file
Call Platform methods in order to perform file system operations
in the AES keylocker detection code. Before the CL, this code was
doing direct file operations with /proc/crypto and
/run/cryptohome/.

This commit is expected to be no-op for production code, but
allows the related code to be tested or fuzzed properly.

BUG=b:254864841
TEST=cros_workon_make cryptohome --test

Change-Id: I7b07979160bb9509605cea2b5f37922dfdd2b3ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4056258
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>