diagnostics: Update Nullable field in |BluetoothDeviceInfo|
This CL updates three fields from NullableXXXX to NullableXXXX? and
passes the mojo pre-commit check. However, when the corresponding struct
field on the receiver side (Chromium) is not optional and the data is
null, it could break mojo backward compatibility.

This CL is safe because:
1. The chromium side has no client to use this struct.
2. Even if the struct is in use, we'll make sure to not send null
   fields in this struct until chromium finishes uprev (CL:4060052).

BUG=b:260364134
TEST=FEATURES=test P2_TEST_FILTER='*' emerge-volteer diagnostics

Change-Id: If6b048c97f6565c70d6c1994017c109423083728
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4061386
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Tested-by: Byron Lee <byronlee@chromium.org>
Commit-Queue: Byron Lee <byronlee@chromium.org>