shill: refactor link monitor event handling
This patch partially moves the handling of patchpanel's link monitor
events to the Network class. This allows to share some of the processing
for link monitor events over WiFi and Ethernet.

The definition of the main callback is moved to the
Network::EventHandler virtual interface. This is part of an effort to
regroup all networking behaviors and features under the Network class.

Semantically the only difference is that the DBus Service
"LinkMonitorDisable property is not considered separately for every new
neighbor reachability event, but only once per network connection when
Network::Start() is called. Since this property is only accessible
through policies, there is no concrete impact because a change in a
policy configured Service that is already connected always causes that
Service to be reconnected. This change eliminates one dependency on
Service.

BUG=b:229309479
TEST=FEATURES=test emerge-$BOARD patchpanel-client shill

Change-Id: Idba6903679bd20cddfcb51c99f09710b889cdf29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4068683
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>