cryptohome: Utilize absl::Cleanup in the fingerprint service code
This isn't critical but after getting cleanup into the allowed list of
absl utilities I wanted to add at least one demonstration of using it.
There were a couple of places in the FP service code where it would've
been handy so I did a bit of refactoring to add it in.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Idd30e98cf2447bbf0b770a7a2774d32aac3831ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4062826
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>