typecd: Add "None" case to WrongConfig metric
Currently, the WrongConfigurationMetric tracks cases where a USB-C
connection is limited by the cable in some way. But, we cannot tell what
percent of the time users are limited by their cable because the metric
does not include a success case. This CL adds a "None" option to the
WrongConfig metric which gets reported when the user is using an
appropriate cable for their device.

BUG=b:260628637
TEST=Passing unit tests. Also appears in
ChromeOS.TypeC.WrongConfiguration histogram in chrome://histograms after
triggering this case by connecting a TBT4 dock with TBT4 cable.

Change-Id: I526ee816ee0b5a7abdeaeab72db5b3d2feabb769
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4062835
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>