power: Avoid sending multiple signals on IncreaseKeyboardBrightness
Each time the DBus API `IncreaseKeyboardBrightness` or
`DecreaseKeyboardBrightness` is called, `powerd` emits a signal
indicating the new brightness. This signal is emitted even if no change
in brightness is made, so that the user gets a visual indication
that their input was processed (even if nothing changed).

However, the current code will emit multiple signals on a transition
from almost-full brightness to full brightness, or almost-off to off. In
particular, the existing code in `IncreaseKeyboardBrightness` and
`DecreaseKeyboardBrightness` tries to guess when a signal won't be sent,
and manually sends the signal itself. However, in the edge cases
described above, the guess is incorrect, and two signals are sent.

Update the code to make `UpdateState` and `ApplyBrightnessPercent`
always responsible for sending the signal. We add a parameter to the two
functions allowing callers to indicate if signals should unconditionally
be sent, ensuring that signals are always sent for user-initiated
actions.

BUG=b:260774697
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`
TEST=Manually increase / decrease brightness in Ash, confirm UI updates.

Change-Id: I82a572de4ab2b8c32d8053bb4a978736c07ff458
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4060299
Tested-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Yanye Li <yanyeli@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: David Lattimore <dml@chromium.org>