power: Handle NaN in ClampPercent
ClampPercent ensures that the input double is a valid value in the range
[0, 100] using a `std::max` and `std::min`.

However, if the input value is a double NaN value, the implementation
triggers undefined behaviour. Currently, this is not a problem, but
upcoming CLs will start using this function to sanitize values coming
from untrusted sources: in these cases, we want to ensure that all
input values are correctly handled.

Update the function to detect NaN values, and return a value in the
range [0.0, 100] as expected. Additionally, add some tests exercising
the function.

BUG=b:254292590
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`

Change-Id: Ibcfc15e641cfa68e2edf4fe3e80666c43c53f98c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4060636
Reviewed-by: David Lattimore <dml@chromium.org>
Tested-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>