uwbd: create an empty project for UWB daemon
This CL creates an new empty project for the UWB daemon. The OWNERS
inside this folder is the root of this project. The OWNERS at ebuild
folder will include it later. So we need to merge this CL first.

BUG=b:259490109
TEST=none

Change-Id: Idbdfc9ccc258c5baf87adb69018da1915670375b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4063090
Reviewed-by: Ningyuan Wang <ningyuan@google.com>
Auto-Submit: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>