libbrillo: Fix static var in LoopDeviceFake
Refactor LoopDeviceFake to not have a statically allocated std::vector
variable. A static variable meant that state was leaking from one
unit test to another, and also meant memory leak for fuzzers that
trigger this code.

It seems this code could've refactored further to make more direct
use of callback parameter binding and state passing and to get rid
of a bunch of reinterpret casts, but I didn't try it in the current
commit.

BUG=b:254864841
TEST=none

Change-Id: I84bb36ebeb30f72b36f7ddd18d998abc7129bbe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4061897
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>