secanomalyd: removes timer based reporting
Changes the anomaly reporting methodology such that it is not based on a
constant timer, rather based on when an anomalous condition is detected.
This is to eliminate the time gap between when the condition is detected
and when it is reported, which causes discrepancies in the metadata
included at each stage due to possible changes in the system state.
UMA metric for W+X mount count is still uploaded every two hours.
Additionally, changes are made to clear the |wx_mounts_| map after every
successful UMA metric upload.

BUG=b:260721141, b:219574442
TEST=All local tests pass.
TEST=secanomalyd scans for anomalies every 30 seconds and invokes crash
reported only when an anomaly is observed.

Change-Id: Ifbe04bc77b527bb44ef18eb97078133504801ad5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4064884
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>