spaced: spaced_cli prints "error" instead of "-1" in case of error
If spaced_cli cannot determine a size (for example if the spaced daemon
is not running), it now prints the string "error" instead of the
negative number -1. This is more explicit and reduces the possilibity of
a negative size to be parsed and interpreted incorrectly down the track.

BUG=b:260805682
TEST=Run spaced_cli

Change-Id: I74cf336b12e7783a18edc023b0f3f8857a3ed532
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4065640
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>