rgbkbd: add eventfd2 and statfs[64] syscalls to the seccomp filters
Newer versions of libusb switched to using eventfd for signalling and
we need to add eventfd2 syscall to the list of allowed ones.

Additionally newer versions of libusb use statfs() to check if sysfs is
available, and this syscall was missing from arm64 and arm device
policies.

BUG=b:247573723
TEST=Boot brya and jacuzzi[64] with libusb-1.0.26 and see that rgbkbd
process does not crash with seccomp violation

Change-Id: I6657e205356e9c2691dbe0c9ff04af4bd218204a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4121308
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>