rgbkbd: add eventfd2 and statfs[64] syscalls to the seccomp filters
Newer versions of libusb switched to using eventfd for signalling and
we need to add eventfd2 syscall to the list of allowed ones.

Additionally newer versions of libusb use statfs() to check if sysfs is
available, and this syscall was missing from arm64 and arm device
policies.

BUG=b:247573723
TEST=Boot brya and corsola with libusb-1.0.26 and see that rgbkbd
process does not crash with seccomp violation

Change-Id: I14f77ba52f36925d0ea031fae34a16aa30fc46da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4068044
Reviewed-by: David Padlipsky <dpad@google.com>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>