ml: Modify SpeechRecognizerEventFromProto to print logs.
Adds support to print logs from ml service, and long with that change
the expected type to be a std::optional , since not all SodaResponse
have a corresponding mojom message to send.

Modify the unit test to support, and downgrade a DFATAL to an ERROR
for unknown types.

BUG=b:257166700
TEST=Runs happily on device and unittests pass.

Change-Id: Ib210f687e82c99ef507e0e3c991d4ddd5818784d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4066442
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Tested-by: Rob Schonberger <robsc@chromium.org>