cryptohome: Store Keylocker in non-static var
Don't store the AES keylocker availability in a static
variable. Static variables are a global state that's
shared between different unit tests or between fuzzer
runs.

Instead, store the availability in  variable in the HomeDirs
object. That way, the check will still happen once when running
in production, but it'll be fully mockable and testable in
tests and fuzzers.

BUG=b:254864841
TEST=cros_workon_make --test cryptohome

Change-Id: Idefb0238c4956944913cdd915cacfa85fe8bafe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4066127
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>