sepolicy: Add cros_health_tool context for test commands
cros-health-tool is one of the clients for cros_healthd, allowing for
terminal level interaction for diagnostics, telemetry and events.
Add sepolicy rules to translate the ssh context and
tty context to the added cros_health_tool, in order to allow the
access of the healthd services in Mojo Service Manager.

BUG=b:230064284
TEST=run on zork from ssh and from crosh

Change-Id: Ib478deb3c92c6973cfdea1d2c513a6022ad0b863
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4066208
Reviewed-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: Ethan Cheng <yycheng@google.com>
Reviewed-by: Qijiang Yūki Ishii <fqj@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Tested-by: Ethan Cheng <yycheng@google.com>