cryptohome: Crypto::Init returns void
This method was always returning true, so there's no point in
returning this boolean and checking it at callsites. Also it
fixes a small bug in the test in which this call was put into
DCHECK, which made it a no-op in Release builds.

As a bit of drive-by cleanup, make the Crypto class non-virtual.

BUG=b:260898859
TEST=cros_workon_make --test cryptohome

Change-Id: I2371f4145ceb592cfacd3e6ae32d3ca1457e80e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4071285
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>