cryptohome: Some AuthFactorMap cleanup
This applies some cleanup changes that were implemented as (part of)
another CL that was abandoned (chromium:4038174). Since these parts were
generally an improvement I extracted them to a separate, standalone CL.

The main change is dropping all of the const vs non-const complexity in
the stored factor view, and the iterator. I originally based this map
off of UserSessionMap, but there are cases where some clients need to
get a non-const UserSession. That's not the case here, so having
complexity to support mutable vs immutable iteration and lookups is not
useful.

It also switches the iterator over to using the stored factor view,
instead of just the factor. In practice nobody needs this as the code
that iterates over all the factors does not currently care about the
backend type, but it makes the overall API more consistent, and is
easier to extend in the future if we do find other data we want to add
alongside the factor.

BUG=b:258531076
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ifa4a799e1ea0b4efa1dadda4b88a6b71801b1222
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4072024
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>