fusebox: remove read/write via file descriptor
Earlier iterations of the fusebox client were able to receive an FD
(file descriptor) over D-Bus from the fusebox server (chrome). While
this improved I/O performance, it was quite complicated to coordinate
across both processes for when the file descriptor was no longer used
(in the client) so that resources could be unpinned (in the server,
chrome, and possibly in further-upstream processes such as mtpd).

Thus, the fusebox server currently does not send FDs, especially as we
have moved from regular D-Bus RPCs to protobuf-over-D-Bus RPCs. In the
client code, none of the fusebox::OpenFile calls currently pass an FD.

Client "read/write via an FD" code is therefore essentially unused code.
This commit deletes it for now (we can always bring it back in future
commits if needed), which facilitates further simplification stemming
from the move from client-generated to server-generated handles.

BUG=b:255520194
TEST=manual

Change-Id: Ieace9044556c1e3bf88a5c9ddbb3d1cf9dab87fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4074565
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>