cryptohome: Make AuthSessionIds stable in tests
Make values of AuthSessionId in tests and fuzzers stable and
predictable. To achieve this, inject code that generates
base::UnguessableToken from a random generator seeded from a
permanent constant.

The main benefit is for fuzzing, because unstable
AuthSessionIds, when inserted into arbitrary places into
randomly-generated protobufs, lead to unreproducible fuzzing:
in some runs a fragment of an AuthSessionId blob can be
interpreted as a protobuf message with some values, in
others - rejected as garbage.

This change is a no-op for production code, which continues
to use standard libchrome's secure way of generating these
values.

BUG=b:254864841
TEST=run userdataauth_fuzzer with same corpus twice, check same D-Bus calls are triggered

Change-Id: I0e32d5501cee71f838d89f742225ee0a96661d0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4075531
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>