shill: common WiFi channel width for RX/TX
The width of the channel is the same for RX and TX, don't duplicate the
width field for RX and TX.
When emitting the structured event, the event itself still has 2 fields,
one for RX and one for TX to avoid breaking backward compatibility. Both
fields of the event will have the same value.

BUG=b:230525092
BUG=b:239864299
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill

Change-Id: I5110504deefa66775a07d612b7f21418f73409cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4075014
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: David Ruth <druth@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>