cryptohome: Migrate password, kiosk and smartcard on authentication
This CL connects the UssMigrator to AuthenticateAuthFactor.
Creation of the reset secret is not included in this CL to keep the
CL focus narrower. The AuthInput from the AuthenticateAuthFactor is
the right AuthInput to add a new AuthFactor if the AuthFactor type
is not PIN. Hence this CL covers all the cases except for the PIN.

BUG=b:258465405
TEST=USE='uss_migration' FEATURES=test emerge-octopus cryptohome
TEST= - Emerge cryptohome with the USe flag, deploy it.
- Disable USS flag and create a user
- Enable USS flag, login and observe that UserSecretStash and
AuthFactors are created.
- Sign out and sign in again
- Remove the vault keyset file and see that login still succeeds.

Change-Id: If6b151cf7d36c02035a61df0080250d3b79edf1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4075019
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>