mtpd: add eventfd2 and statfs syscalls to the seccomp filters
Newer versions of libusb switched to using eventfd for signalling and
we need to add eventfd2 syscall to the list of allowed ones.

Additionally newer versions of libusb use statfs() to check if sysfs is
available, and this syscall was missing from arm64 and arm device
policies.

BUG=b:247573723
TEST=Boot brya and corsola with libusb-1.0.26 and see that mtpd
process does not crash with seccomp violation

Change-Id: I868f7fa3e1a360049db8ac1f713eb45c55796e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4076844
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: Dmitry Torokhov <dtor@chromium.org>