power: Ensure `keyboard_backlight_turn_on_for_user_activity` set
Ensure the flag `keyboard_backlight_turn_on_for_user_activity` is not
overridden to "false" when powerd daemon starts up.

Investigations into the history of the flag suggest that the flag has
always been set to "true" ever since the feature was first introduced.
b/260034799 investigates removing the flag, allowing us to clean up the
code that implements the old behaviour.

As a final step prior to completely removing the flag, this CL adds
a `CHECK` at powerd startup to ensure the flag is the expected value.
Any unexpected locations where the flag is overridden to "false" should
become loud and clear.

As an implementation note, we check the flag in daemon setup instead of
the keyboard backlight code itself because unit tests still exercise the
"false" path, and we don't want to remove the tests until the underlying
code has also been removed.

BUG=b:260034799
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`

Change-Id: Ica379c09e9ea0ca1fb891a28aab03f52eca12ab6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4075000
Tested-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Mengqi Guo <mqg@chromium.org>