cryptohome: Fix EndFingerprintAuthSession crash
EndFingerprintAuthSession shouldn't crash when there's no
fingerprint manager.

This is a legacy API that'll be removed in a couple of
releases, still it makes sense to fix it for the remaining
time and also to allow userdataauth_fuzzer to explore other
datapaths besides this "quick" crash.

BUG=b:261444565
TEST=new unit test (cros_workon_make --test cryptohome)

Change-Id: I4726dfea57e344fcb641fde9d4cad6172b275838
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4075693
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>