shill: cellular: ensure bearer during link establishment
The link establishment procedure requires a valid bearer object, so
ensure one is available and in the unlikely case there is none, just
trigger an explicit disconnection.

The unit tests are updated to avoid running EstablishLink() on the
Connect() tests, so that they don't fail because an active bearer
isn't set in the capability. The EstablishLink() method is tested
separately for the different IP config methods reported by
ModemManager.

BUG=b:262290710
TEST=FEATURES=test emerge-${BOARD} shill

Change-Id: I41ba044276089a2a29381d26e0e9cf6cb3a38ee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4079039
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>