shill: cellular: validate data interface during connection attempt
The network interface associated to the Cellular device was selected
before creating the object by looking at which net port was listed in
the 'Ports' property in the ModemManager.Modem interface.

The network interface to be used during a connection is reported by
ModemManager as data interface in the bearer info obtained when the
attempt is successful.

In most cases these two network interfaces will be the same, and
therefore the connection will succeed. But in certain cases, these
interfaces will be different (e.g. if the modem has more than one
physical network interface, or if multiplexing has been enabled).

For now, explicitly disconnect the modem if MM asks to connect a
different network interface.

BUG=b:262290710
TEST=FEATURES=test emerge-${BOARD} shill

Cq-Depend: chromium:4096799
Change-Id: I7d76905e271ba46b913d58f4eb52d6b7bf8985da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4079040
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>