power: return an error if display battery percentage doesn't make sense
We're seeing incorrect 0 values for the display battery percentage due
to I2C errors. This performs a check on the two values if display
battery percentage is read from the EC to make sure there isn't a
discrepency. An old workaround for this is removed. That workaround
usually wouldn't help, since it would reread the value from the EC. The
EC only updates the battery values every minute or so, which causes that
fix to not work.

BUG=b:261044627
TEST=unittests
TEST=power_LoadTest and check that "Display and battery..." doesn't show
up in the powerd.LATEST log

Change-Id: I4c87288a845fcd80abb1a67ac6f2aee4cb9c2a0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4081904
Reviewed-by: Alan Green <avg@chromium.org>
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>