dlcservice: Update installed meaning for GetInstalled
This is so clients/tests that depend on DLC being installed get back
only installed AND mounted DLCs. (Instead of adding a new API)
The previous definition of installed being either installed OR mounted
was a bit loose and caused issues when developers depend on the list of
"installed" DLCs.

BUG=none
TEST=FEATURES=test emerge-$B dlcservice
TEST=deploy

Change-Id: I7092ed49c1789ee8bdd197871dd949fd01c7d21e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4080822
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>