crostini: return container_token via ContainerStartedSignal
Cicerone already has the container token when a container is started,
but it is not returned in ContainerStartedSignal. This CL adds a new
field to the signal to include the token. This token will be used to
identify the container's app windows later as described in
go/guestos-shelf-icons.

BUG=b:244651040
TEST=CQ

Change-Id: I633da4ca11fbe87d7d1cd212d0fb286297b5db7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4082791
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Tested-by: Sophia Lin <sophialin@google.com>