shill: Destroy current Connection on DHCP failure
Network::OnDHCPFailure() can be triggered by a failure for DHCP
renewal. When it happens, it means that currently we are holding an
invalid DHCP lease, and thus should remove the IP address and routes
installed in kernel by this lease. This patch adds this logic by
resetting the Connection object before reinstalling it with IPv6 config.
This patch will not affect the logic in the case that there is
StaticIPConfig installed for this network.

Note that ideally, we should reset the IPv4 connection right after
DHCP lease is expired, while currently after this patch, it happens
after the next DHCP negotiation timed out (30 seconds after lease
expired).

BUG=b:261681299
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=tast run $DUT network.Routing*, including the new added
  RoutingDualStackLoseDHCP for this case.

Change-Id: I683a8bf96c043fc4e23466170e936f0b31fd5804
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4086081
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>