cryptohome: Add a metric USS migration results
This adds a new metric where success or errors from USS migration can be
reported. The current version of this so far is just a stub as we don't
yet have any of code hooking up the migration to authssions that would
call this.

BUG=b:258711982
TEST=built, currently no calls to test

Change-Id: I778598f6322f61f56d1d9b87e59fc22bfede8729
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4086040
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>