camera: usb: Add kQuirkPreferLargePreviewResolution to Rapoo C260 webcam
When switching to photo resolution to take a picture, the camera adjusts
color tone in the first several frames and causes bad image quality. To
avoid this, add the quirk to start V4L2 stream with photo resolution so
it doesn't need to switch the resolution.

BUG=b:245461023
TEST=Take pictures in CCA.

Change-Id: I65c50d42fa601afaae7edad9c14713bc723aecd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4087724
Reviewed-by: Danny Kuo <danny_kuo@compal.corp-partner.google.com>
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Danny Kuo <danny_kuo@compal.corp-partner.google.com>
Commit-Queue: Ren-Pei Zeng <kamesan@chromium.org>